+++ src/lib.rs
@@ -6,7 +6,9 @@
 pub mod action_gateway;
 pub mod auth;
 pub mod backend;
+pub mod cloudevents;
 pub mod config;
+pub mod filter;
 pub mod grpc;
 pub mod http_backend;
 pub mod log_buffer;
@@ -16,6 +18,7 @@ pub mod model;
 pub mod push;
 pub mod relay_events_backend;
 pub mod rest;
+pub mod schema_validate;
 pub mod tls;
 
 pub mod google {